refactor: consolidate value key and utility functions into utils module
